+++ base-react/navigation/link/index.ts
@@ -3,5 +3,5 @@ export type { LinkProps, LinkType, Location } from './link.type';
 export { useLocation } from './use-location';
 
 // rename file
-export { useRouter, RouterProvider } from './router-context';
-export type { RouterContextType } from './router-context';
+export { useRouter, RouterProvider } from './routing-provider';
+export type { RouterContextType } from './routing-provider';
